==> scripts/0_intercoder_reliability_check.ipynb <==
import pandas as pd

from sklearn.metrics import accuracy_score, cohen_kappa_score, confusion_matrix

IRR_df = pd.read_csv('./../data/intercoder_reliability_check.csv')

IRR_df.head()

%%capture output

#TABLE A1

print(len(IRR_df[IRR_df['target']=='wm']))

print(accuracy_score(IRR_df['sentiment_C1'][IRR_df['target']=='wm'], IRR_df['sentiment_C2'][IRR_df['target']=='wm']))
print(cohen_kappa_score(IRR_df['sentiment_C1'][IRR_df['target']=='wm'], IRR_df['sentiment_C2'][IRR_df['target']=='wm']))

print(confusion_matrix(IRR_df['sentiment_C1'][IRR_df['target']=='wm'], IRR_df['sentiment_C2'][IRR_df['target']=='wm']))

output()

with open('./../tables/Table_A1.txt', 'w') as f:
    f.write(output.stdout)

%%capture output

#TABLE A2

print(len(IRR_df[IRR_df['target']=='wm']))

print(accuracy_score(IRR_df['stance_C1'][IRR_df['target']=='wm'], IRR_df['stance_C2'][IRR_df['target']=='wm']))
print(cohen_kappa_score(IRR_df['stance_C1'][IRR_df['target']=='wm'], IRR_df['stance_C2'][IRR_df['target']=='wm']))

print(confusion_matrix(IRR_df['stance_C1'][IRR_df['target']=='wm'], IRR_df['stance_C2'][IRR_df['target']=='wm']))

output()

with open('./../tables/Table_A2.txt', 'w') as f:
    f.write(output.stdout)

%%capture output

#TABLE A3

print(len(IRR_df[IRR_df['target']=='kav']))

print(accuracy_score(IRR_df['sentiment_C1'][IRR_df['target']=='kav'], IRR_df['sentiment_C2'][IRR_df['target']=='kav']))

print(cohen_kappa_score(IRR_df['sentiment_C1'][IRR_df['target']=='kav'], IRR_df['sentiment_C2'][IRR_df['target']=='kav']))


print(confusion_matrix(IRR_df['sentiment_C1'][IRR_df['target']=='kav'], IRR_df['sentiment_C2'][IRR_df['target']=='kav']))

output()

with open('./../tables/Table_A3.txt', 'w') as f:
    f.write(output.stdout)

%%capture output

#TABLE A4

print(len(IRR_df[IRR_df['target']=='kav']))

print(accuracy_score(IRR_df['stance_C1'][IRR_df['target']=='kav'], IRR_df['stance_C2'][IRR_df['target']=='kav']))
print(cohen_kappa_score(IRR_df['stance_C1'][IRR_df['target']=='kav'], IRR_df['stance_C2'][IRR_df['target']=='kav']))

print(confusion_matrix(IRR_df['stance_C1'][IRR_df['target']=='kav'], IRR_df['stance_C2'][IRR_df['target']=='kav']))

output()

with open('./../tables/Table_A4.txt', 'w') as f:
    f.write(output.stdout)
